refactor: lift parseStrongEtag to common and move blob cleanup to IO
- parseStrongEtag was duplicated in ModuleRoute and BlobRoute; lift it to HttpExtensions as a shared top-level helper.

- commitModule's orphan blob cleanup launched deleteRecursively on the default dispatcher; switch to Dispatchers.IO to match deleteBlob and free the Default pool from blocking FS work.
