+++ src/main/kotlin/eu/darken/octi/server/module/ModuleRoute.kt
@@ -7,6 +7,7 @@ import eu.darken.octi.server.common.debug.logging.asLog
 import eu.darken.octi.server.common.debug.logging.log
 import eu.darken.octi.server.common.debug.logging.logTag
 import eu.darken.octi.server.common.debug.logging.shortId
+import eu.darken.octi.server.common.parseStrongEtag
 import eu.darken.octi.server.common.verifyCaller
 import eu.darken.octi.server.device.Device
 import eu.darken.octi.server.device.DeviceId
@@ -283,23 +284,6 @@ class ModuleRoute @Inject constructor(
             moduleId = moduleId,
             action = "updated",
         )
-    }
-
-    /**
-     * Parses an entity-tag for `If-Match` / `If-None-Match`.
-     * Accepts `*`, `"opaque"`, and bare `opaque` (legacy clients).
-     * Rejects weak (`W/"..."`) — If-Match requires strong comparison (RFC 7232 §3.1).
-     * Returns null for malformed input so the caller can respond 400.
-     */
-    private fun parseStrongEtag(raw: String): String? {
-        val trimmed = raw.trim()
-        if (trimmed.isEmpty()) return null
-        if (trimmed == "*") return "*"
-        if (trimmed.startsWith("W/", ignoreCase = true)) return null
-        if (trimmed.startsWith("\"") && trimmed.endsWith("\"") && trimmed.length >= 2) {
-            return trimmed.substring(1, trimmed.length - 1)
-        }
-        return trimmed
     }
 
     companion object {
